+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - support deleting files from a component
 - fix the internal ci-update command where an environment has a tester without a compiler 
 - improved commit, add, export and status outputs
+- support general failures on bit test (like on before)
 - improved bit help
 - fix bit config command for linux
 - status output with missing dependencies 
@@ -26,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - improve outputs
 - improve windows support for import command
 - exception when using `bit test` or `bit build` before adding first components
+- add new flag to bit add to override or append files to bit component
 
 
 ## [0.10.5] - 2017-08-16